Move plugin module into tools directory
