+++ tools/sds-aspect-model-maven-plugin/src/main/java/io/openmanufacturing/sds/aspectmodel/GenerateStaticJavaClasses.java
