+++ tools/sds-aspect-model-maven-plugin/src/test/java/io/openmanufacturing/sds/aspectmodel/GenerateStaticJavaClassesTest.java
